+++ tools/perf/tests/switch-tracking.c
@@ -2,6 +2,7 @@
 #include <sys/time.h>
 #include <sys/prctl.h>
 #include <errno.h>
+#include <inttypes.h>
 #include <limits.h>
 #include <time.h>
 #include <stdlib.h>
@@ -487,7 +488,7 @@ static int test__switch_tracking(struct test_suite *test __maybe_unused, int sub
 	}
 
 	comm = "Test COMM 1";
-	err = prctl(PR_SET_NAME, (unsigned long)comm, 0, 0, 0);
+	err = prctl(PR_SET_NAME, (uintptr_t)comm, 0, 0, 0);
 	if (err) {
 		pr_debug("PR_SET_NAME failed!\n");
 		goto out_err;
@@ -500,7 +501,7 @@ static int test__switch_tracking(struct test_suite *test __maybe_unused, int sub
 	}
 
 	comm = "Test COMM 2";
-	err = prctl(PR_SET_NAME, (unsigned long)comm, 0, 0, 0);
+	err = prctl(PR_SET_NAME, (uintptr_t)comm, 0, 0, 0);
 	if (err) {
 		pr_debug("PR_SET_NAME failed!\n");
 		goto out_err;
@@ -513,7 +514,7 @@ static int test__switch_tracking(struct test_suite *test __maybe_unused, int sub
 	}
 
 	comm = "Test COMM 3";
-	err = prctl(PR_SET_NAME, (unsigned long)comm, 0, 0, 0);
+	err = prctl(PR_SET_NAME, (uintptr_t)comm, 0, 0, 0);
 	if (err) {
 		pr_debug("PR_SET_NAME failed!\n");
 		goto out_err;
@@ -526,7 +527,7 @@ static int test__switch_tracking(struct test_suite *test __maybe_unused, int sub
 	}
 
 	comm = "Test COMM 4";
-	err = prctl(PR_SET_NAME, (unsigned long)comm, 0, 0, 0);
+	err = prctl(PR_SET_NAME, (uintptr_t)comm, 0, 0, 0);
 	if (err) {
 		pr_debug("PR_SET_NAME failed!\n");
 		goto out_err;
